HIndent

Description

Haskell indenter.

Formatting functions.

reformat :: Style -> Maybe [Extension] -> Text -> Either String Builder

Format the given source.

prettyPrint :: Style -> (forall s. Printer s ()) -> Builder

Pretty print the given printable thing.

parseMode :: ParseMode

Parse mode, includes all extensions, doesn't assume any fixities.

Style

data Style

A printer style.

Constructors

forall s . Style

Fields

styleName :: !Text

Name of the style, used in the commandline interface.

styleAuthor :: !Text

Author of the printer (as opposed to the author of the style).

styleDescription :: !Text

Description of the style.

styleInitialState :: !s

User state, if needed.

styleExtenders :: ![Extender s]

Extenders to the printer.

styleDefConfig :: !Config

Default config to use for this style.

styles :: [Style]

Styles list, useful for programmatically choosing.

chrisDone :: Style

The printer style.

johanTibell :: Style

The printer style.

fundamental :: Style

The printer style.

gibiansky :: Style

The printer style.

Testing

test :: Style -> Text -> IO ()

Test with the given style, prints to stdout.

testAll :: Text -> IO ()

Test with all styles, prints to stdout.

testAst :: Text -> Either String ([ComInfo], Module NodeInfo)

Parse the source and annotate it with comments, yielding the resulting AST.