HIndent.Pretty

Description

Pretty printing.

Printing

class (Annotated ast, Typeable ast) => Pretty ast

Pretty printing class.

Minimal complete definition

prettyInternal

pretty :: Pretty ast => ast NodeInfo -> Printer ()

Pretty print using extenders.

prettyNoExt :: Pretty ast => ast NodeInfo -> Printer ()

Run the basic printer for the given node without calling an extension hook for this node, but do allow extender hooks in child nodes. Also auto-inserts comments.

Insertion

write :: Builder -> Printer ()

Write out a string, updating the current position information.

newline :: Printer ()

Output a newline.

space :: Printer ()

Write a space.

comma :: Printer ()

Write a comma.

int :: Integral n => n -> Printer ()

Write an integral.

string :: String -> Printer ()

Write a string.

Common node types

maybeCtx :: Maybe (Context NodeInfo) -> Printer ()

Maybe render a class context.

printComment :: ComInfo -> Printer ()

Pretty print a comment.

Interspersing

inter :: Printer () -> [Printer ()] -> Printer ()

Print all the printers separated by sep.

spaced :: [Printer ()] -> Printer ()

Print all the printers separated by spaces.

lined :: [Printer ()] -> Printer ()

Print all the printers separated by spaces.

prefixedLined :: Text -> [Printer ()] -> Printer ()

Print all the printers separated newlines and optionally a line prefix.

commas :: [Printer ()] -> Printer ()

Print all the printers separated by commas.

Wrapping

parens :: Printer a -> Printer a

Wrap in parens.

brackets :: Printer a -> Printer a

Wrap in brackets.

braces :: Printer a -> Printer a

Wrap in braces.

Indentation

indented :: Int64 -> Printer a -> Printer a

Increase indentation level by n spaces for the given printer.

column :: Int64 -> Printer a -> Printer a

Set the (newline-) indent level to the given column for the given printer.

depend :: Printer () -> Printer b -> Printer b

Make the latter's indentation depend upon the end column of the former.

dependBind :: Printer a -> (a -> Printer b) -> Printer b

Make the latter's indentation depend upon the end column of the former.

swing :: Printer () -> Printer b -> Printer b

Swing the second printer below and indented with respect to the first.

getIndentSpaces :: Printer Int64

Indent spaces, e.g. 2.

getColumnLimit :: Printer Int64

Column limit, e.g. 80

Predicates

nullBinds :: Binds NodeInfo -> Bool

No binds?

Sandboxing

sandbox :: MonadState s m => m a -> m (a, s)

Play with a printer and then restore the state to what it was before.

Fallback

pretty' :: (Pretty ast, Pretty (ast SrcSpanInfo), Functor ast) => ast NodeInfo -> Printer ()

Pretty print using HSE's own printer. The Pretty class here is HSE's.