HIndent.Styles.ChrisDone

Description

Chris Done's style.

Documented here: https://github.com/chrisdone/haskell-style-guide

Documentation

shortName :: Int64

A short function name.

smallColumnLimit :: Int64

Column limit: 50

data State

Empty state.

Constructors

State

chrisDone :: Style

The printer style.

decl :: Decl NodeInfo -> Printer s ()

Pretty print type signatures like

foo :: (Show x,Read x) => (Foo -> Bar) -> Maybe Int -> (Char -> X -> Y) -> IO ()

fieldupdate :: FieldUpdate NodeInfo -> Printer t ()

I want field updates to be dependent or newline.

rhs :: Rhs NodeInfo -> Printer t ()

Right-hand sides are dependent.

unguardedrhs :: Rhs NodeInfo -> Printer t ()

Right-hand sides are dependent.

unguardedalt :: Rhs NodeInfo -> Printer t ()

Unguarded case alts.

contextualGuardedRhs :: GuardedRhs NodeInfo -> Printer t ()

Decide whether to do alts or rhs based on the context.

guardedrhs :: GuardedRhs NodeInfo -> Printer t ()

I want guarded RHS be dependent or newline.

guardedalt :: GuardedRhs NodeInfo -> Printer t ()

I want guarded alts be dependent or newline.

stmt :: Stmt NodeInfo -> Printer t ()

exp :: Exp NodeInfo -> Printer t ()

Expressions

sandboxSingles :: Pretty ast => [ast NodeInfo] -> Printer t (Bool, PrintState t)

Sandbox and render the nodes on multiple lines, returning whether each is a single line.

multi :: Pretty ast => Int64 -> [ast NodeInfo] -> Bool -> Printer t ()

Render multi-line nodes.

sandboxNonOverflowing :: Pretty ast => [ast NodeInfo] -> Printer t ((Bool, Bool), PrintState t)

Sandbox and render the node on a single line, return whether it's on a single line and whether it's overflowing.

isShort :: Pretty ast => ast NodeInfo -> Printer t (Bool, PrintState t)

Is the expression "short"? Used for app heads.

isSmall :: MonadState (PrintState t) m => m a -> m (Bool, PrintState t)

Is the given expression "small"? I.e. does it fit on one line and under smallColumnLimit columns.

isFlat :: Exp NodeInfo -> Bool

Is an expression flat?

isOverflow :: Printer t a -> Printer t Bool

Does printing the given thing overflow column limit? (e.g. 80)

isOverflowMax :: Printer t a -> Printer t Bool

Does printing the given thing overflow column limit? (e.g. 80)

isSingleLiner :: MonadState (PrintState t) m => m a -> m Bool

Is the given expression a single-liner when printed?

infixApp :: (Pretty ast, Pretty ast1, Pretty ast2) => Exp NodeInfo -> ast NodeInfo -> ast1 NodeInfo -> ast2 NodeInfo -> Maybe Int64 -> Printer t ()

dependOrNewline :: Printer t () -> Exp NodeInfo -> (Exp NodeInfo -> Printer t ()) -> Printer t ()

Make the right hand side dependent if it's flat, otherwise newline it.