HIndent.Styles.Gibiansky

Documentation

data State

Empty state.

Constructors

State

Fields

gibianskyForceSingleLine :: Bool

gibiansky :: Style

The printer style.

indentSpaces :: Integral a => a

Number of spaces to indent by.

indentOnce :: Printer s ()

Printer to indent one level.

maxSingleLineExports :: Integral a => a

How many exports to format in a single line. If an export list has more than this, it will be formatted as multiple lines.

attemptSingleLine :: Printer State a -> Printer State a -> Printer State a

type Extend f = f NodeInfo -> Printer State ()

modl :: Extend Module

Format whole modules.

pragmas :: Extend ModulePragma

Format pragmas differently (language pragmas).

pat :: Extend Pat

Format patterns.

imp :: Extend ImportDecl

Format import statements.

context :: Extend Context

Format contexts with spaces and commas between class constraints.

derivings :: Extend Deriving

Format deriving clauses with spaces and commas between class constraints.

typ :: Extend Type

Format function type declarations.

writeTuple :: Pretty ast => Boxed -> [ast NodeInfo] -> Printer State ()

sameLine :: (Annotated ast, Annotated ast') => ast NodeInfo -> ast' NodeInfo -> Bool

collectTypes :: Type l -> [Type l]

exprs :: Extend Exp

letExpr :: Exp NodeInfo -> Printer State ()

appExpr :: Exp NodeInfo -> Printer State ()

doExpr :: Exp NodeInfo -> Printer State ()

listExpr :: Exp NodeInfo -> Printer State ()

singleLineList :: [Exp NodeInfo] -> Printer State ()

multiLineList :: [Exp NodeInfo] -> Printer State ()

dollarExpr :: Exp NodeInfo -> Printer State ()

applicativeExpr :: Exp NodeInfo -> Printer State ()

opExpr :: Exp NodeInfo -> Printer State ()

lambdaExpr :: Exp NodeInfo -> Printer State ()

caseExpr :: Exp NodeInfo -> Printer State ()

lambdaCaseExpr :: Exp NodeInfo -> Printer State ()

ifExpr :: Exp NodeInfo -> Printer State ()

writeCaseAlts :: [Alt NodeInfo] -> Printer State ()

recUpdateExpr :: Printer State () -> [FieldUpdate NodeInfo] -> Printer State ()

rhss :: Extend Rhs

guardedRhs :: Extend GuardedRhs

rhsRest :: Pretty ast => ast NodeInfo -> Printer State ()

decls :: Extend Decl

funBody :: [Pat NodeInfo] -> Rhs NodeInfo -> Maybe (Binds NodeInfo) -> Printer State ()

writeWhereBinds :: Binds NodeInfo -> Printer State ()

onSeparateLines :: (Pretty ast, Annotated ast) => [ast NodeInfo] -> Printer State ()

astStartLine :: Annotated ast => ast NodeInfo -> Int

isDoBlock :: Rhs l -> Bool

condecls :: Extend ConDecl

alt :: Extend Alt

moduleHead :: Extend ModuleHead

exportList :: Extend ExportSpecList

lineDelta :: (Annotated ast1, Annotated ast2) => ast1 NodeInfo -> ast2 NodeInfo -> Int

fieldUpdate :: Extend FieldUpdate