HIndent.Styles.JohanTibell

Description

Stub module for Johan Tibell's style.

Documented here: https://github.com/tibbe/haskell-style-guide/blob/master/haskell-style.md

Questions:

How to indent after a guarded alt/rhs? How to indent let? How to indent large ADT constructors types?

Documentation

shortName :: Int64

A short function name.

data State

Empty state.

Constructors

State

johanTibell :: Style

The printer style.

stmt :: Stmt NodeInfo -> Printer s ()

rhs :: Rhs NodeInfo -> Printer s ()

Handle do specially and also space out guards more.

guardedRhs :: GuardedRhs NodeInfo -> Printer s ()

Handle do specially.

Implement dangling right-hand-sides.

unguardedalt :: Rhs NodeInfo -> Printer s ()

Unguarded case alts.

unguardedrhs :: Rhs NodeInfo -> Printer s ()

exp :: Exp NodeInfo -> Printer s ()

Space out tuples.

Expression customizations.

decl :: Decl NodeInfo -> Printer s ()

Pretty print type signatures like

foo :: (Show x,Read x) => (Foo -> Bar) -> Maybe Int -> (Char -> X -> Y) -> IO ()

Specially format records. Indent where clauses only 2 spaces.

qualConDecl :: QualConDecl NodeInfo -> Printer s ()

Use special record display, used by dataDecl in a record scenario.

conDecl :: ConDecl NodeInfo -> Printer s ()

Fields are preceded with a space.

recDecl :: ConDecl NodeInfo -> Printer s ()

Record decls are formatted like: Foo { bar :: X }

recUpdateExpr :: Printer s () -> [FieldUpdate NodeInfo] -> Printer s ()

isRecord :: QualConDecl t -> Bool

Is the decl a record?

isOverflow :: Printer s a -> Printer s Bool

Does printing the given thing overflow column limit? (e.g. 80)

isSingleLiner :: MonadState (PrintState s) m => m a -> m Bool

Is the given expression a single-liner when printed?

isShort :: Pretty ast => ast NodeInfo -> Printer s Bool

Is the expression "short"? Used for app heads.

isFlat :: Exp NodeInfo -> Printer s Bool

Is an expression flat?

fieldupdate :: FieldUpdate NodeInfo -> Printer s ()

rhs on field update on the same line as lhs.

infixApp :: (Pretty ast, Pretty ast1, Pretty ast2) => Exp NodeInfo -> ast NodeInfo -> ast1 NodeInfo -> ast2 NodeInfo -> Maybe Int64 -> Printer s ()