HIndent.Pretty

Description

Pretty printing.

Printing

class (Annotated ast, Typeable ast) => Pretty ast

Pretty printing class.

Minimal complete definition

prettyInternal

pretty :: (Pretty ast, MonadState (PrintState s) m) => ast NodeInfo -> m ()

Pretty print using extenders.

prettyNoExt :: (Pretty ast, MonadState (PrintState s) m) => ast NodeInfo -> m ()

Run the basic printer for the given node without calling an extension hook for this node, but do allow extender hooks in child nodes. Also auto-inserts comments.

User state

getState :: Printer s s

Get the user state.

putState :: s -> Printer s ()

Put the user state.

modifyState :: (s -> s) -> Printer s ()

Modify the user state.

Insertion

write :: MonadState (PrintState s) m => Builder -> m ()

Write out a string, updating the current position information.

newline :: MonadState (PrintState s) m => m ()

Output a newline.

space :: MonadState (PrintState s) m => m ()

Write a space.

comma :: MonadState (PrintState s) m => m ()

Write a comma.

int :: (Integral n, MonadState (PrintState s) m) => n -> m ()

Write an integral.

string :: MonadState (PrintState s) m => String -> m ()

Write a string.

Common node types

maybeCtx :: MonadState (PrintState s) m => Maybe (Context NodeInfo) -> m ()

Maybe render a class context.

printComment :: MonadState (PrintState s) m => Maybe SrcSpan -> ComInfo -> m ()

Pretty print a comment.

printComments :: (Pretty ast, MonadState (PrintState s) m) => ComInfoLocation -> ast NodeInfo -> m ()

Print comments of a node.

withCaseContext :: MonadState (PrintState s) m => Bool -> m a -> m a

Set the context to a case context, where RHS is printed with -> .

rhsSeparator :: MonadState (PrintState s) m => m ()

Get the current RHS separator, either = or -> .

Interspersing

inter :: MonadState (PrintState s) m => m () -> [m ()] -> m ()

Print all the printers separated by sep.

spaced :: MonadState (PrintState s) m => [m ()] -> m ()

Print all the printers separated by spaces.

lined :: MonadState (PrintState s) m => [m ()] -> m ()

Print all the printers separated by newlines.

prefixedLined :: MonadState (PrintState s) m => Text -> [m ()] -> m ()

Print all the printers separated newlines and optionally a line prefix.

commas :: MonadState (PrintState s) m => [m ()] -> m ()

Print all the printers separated by commas.

Wrapping

parens :: MonadState (PrintState s) m => m a -> m a

Wrap in parens.

brackets :: MonadState (PrintState s) m => m a -> m a

Wrap in brackets.

braces :: MonadState (PrintState s) m => m a -> m a

Wrap in braces.

Indentation

indented :: MonadState (PrintState s) m => Int64 -> m a -> m a

Increase indentation level by n spaces for the given printer.

column :: MonadState (PrintState s) m => Int64 -> m a -> m a

Set the (newline-) indent level to the given column for the given printer.

getColumn :: MonadState (PrintState s) m => m Int64

Get the current indent level.

getLineNum :: MonadState (PrintState s) m => m Int64

Get the current line number.

depend :: MonadState (PrintState s) m => m () -> m b -> m b

Make the latter's indentation depend upon the end column of the former.

dependBind :: MonadState (PrintState s) m => m a -> (a -> m b) -> m b

Make the latter's indentation depend upon the end column of the former.

swing :: MonadState (PrintState s) m => m () -> m b -> m b

Swing the second printer below and indented with respect to the first.

getIndentSpaces :: MonadState (PrintState s) m => m Int64

Indent spaces, e.g. 2.

getColumnLimit :: MonadState (PrintState s) m => m Int64

Column limit, e.g. 80

Predicates

nullBinds :: Binds NodeInfo -> Bool

No binds?

Sandboxing

sandbox :: MonadState s m => m a -> m (a, s)

Play with a printer and then restore the state to what it was before.

Fallback

pretty' :: (Pretty ast, Pretty (ast SrcSpanInfo), Functor ast, MonadState (PrintState s) m) => ast NodeInfo -> m ()

Pretty print using HSE's own printer. The Pretty class here is HSE's.